+++ core/telegram/client_management.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Telegram客户端管理模块
 """